requirements.txt
click>=8.1.0
gitpython>=3.1.0
rich>=13.0.0
